requirements.txt
fastapi
uvicorn
pydantic
langchain-ollama
langchain-core
mcp
python-dotenv
fastmcp
httpx
